+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased][unreleased]
 ### Added
+- Definition `enable` to enable/disable a Docker Compose command.
 - Definitions for options of Docker Compose commands `build`, `rm`, `scale` and `up`.
 - Definition `verbose` for showing more output in Docker Compose.
 - Definitions for Docker Compose commands `build`, `kill`, `logs`, `port`, `ps`, `pull`, `rm`, `scale`, `start`,